+++ ripper/arg_parser.py
@@ -34,7 +34,10 @@ def parser_add_options(parser: OptionParser) -> None:
                       help=f'Max random packets length. Default: {ARGS_DEFAULT_MAX_RND_PACKET_LEN} for udp/tcp')
     parser.add_option('-y', '--proxy_list',
                       dest='proxy_list',
-                      help='File with sock5 proxies in ip:port:username:password line format. Proxies will be ignored in udp attack!')
+                      help='File (fs or http/https) with proxies in ip:port:username:password line format. Proxies will be ignored in udp attack!')
+    parser.add_option('-k', '--proxy_type',
+                      dest='proxy_type', type='str', default='socks5',
+                      help='Type of proxy to work with. Supported types: socks5, socks4, http. Default: socks5')
     parser.add_option('-c', '--health_check',
                       dest='health_check', type='int', default=ARGS_DEFAULT_HEALTH_CHECK,
                       help=f'Controls health check availability. Turn on: 1, turn off: 0. Default: {ARGS_DEFAULT_HEALTH_CHECK}')