+++ ripper/context.py
@@ -309,10 +309,16 @@ class Context:
         self.cpu_count = max(os.cpu_count(), 1)  # to avoid situation when vCPU might be 0
 
         # Get required data from files
-        self.user_agents = strip_lines(common.readfile(os.path.dirname(__file__) + '/useragents.txt'))
-        self.base_headers = strip_lines(common.readfile(os.path.dirname(__file__) + '/headers.txt'))
+        self.user_agents = strip_lines(common.read_file_lines_fs(os.path.dirname(__file__) + '/useragents.txt'))
+        self.base_headers = strip_lines(common.read_file_lines_fs(os.path.dirname(__file__) + '/headers.txt'))
         self.headers = get_headers_dict(self.base_headers)
 
+    if args[0].proxy_list and _ctx.attack_method != 'udp':
+        _ctx.proxy_manager.set_proxy_type(args[0].proxy_type)
+        try:
+            _ctx.proxy_manager.update_proxy_list_from_file(args[0].proxy_list)
+        except Exception as e:
+             _ctx.add_error(Errors('Proxy list read operation failed', e))
         # Get initial info from external services
         self.IpInfo.my_start_ip = common.get_current_ip()
         self.IpInfo.my_current_ip = self.IpInfo.my_start_ip